+++ CHANGELOG.md
@@ -107,6 +107,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Rename `init_cmd`→`init_cmds` parameter in `[format:deb]` and `[format:rpm]`
     sections of system configuration.
   - Bump Fedora release from 38 to 40 in rpm and osi container images.
+  - Install all mkosi 22 optional dependencies required to get all features and
+    assign newuidmap/newgidmap ranges for Fatbuildr system user in osi container
+    image.
   - Install podman for Mock in rpm container image.
 - docs:
   - Update table of available remote features with new registry content